Adds an angular $formatter to handle model value updates.
Fixes improper formatting of numbers when fields load with an existing ng-model value. This may not be the best way to handle data formatting since it has to change the value in the field in order to return the new value to put in the field, but intl-tel-input does not offer a static method to format a number outside of a field.